pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.11.13",
    "discord-py>=2.5.0",
    "python-dotenv>=1.0.1",
    "trafilatura>=2.0.0",
]
